refactor sagemaker. we want to split base and sagemaker implementation classes
